# fix: Memory leak in watch mode and use Set for performance