## TLDR
Before: *200ms* per rebuild (and an additional 200ms for each additional rebuild) 
After: *~5ms* per rebuild!

## The Long Version
Hi! Thanks for the awesome plugin! It has saved my life on many projects. One of my coworkers today was experiencing slow rebuilds in watch mode, and with a little bit of profiling, it turns out this plugin was the culprit. There are 2 issues:

1. a memory leak that causes the fileDependencies and contextDependencies to grow after each save.
2. An `O(n)` `_.includes` call that can be replaced with an `O(1)` `Set.prototype.has` call.

Here are the performance characteristics of this patch. The project in question is private, but has 1800 files. The `did it` line is how long this plugin takes to complete:

## Before Fixing Memory Leak

```
did it: 386.245ms                                                     
Hash: fd31795dd4d692c70ccd
Version: webpack 2.3.3
Time: 2522ms

did it: 597.102ms                                                     
Hash: 77299979c1a9afc21683
Version: webpack 2.3.3
Time: 2863ms

did it: 827.018ms                                                     
Hash: fd31795dd4d692c70ccd
Version: webpack 2.3.3
Time: 2959ms

did it: 1078.640ms                                                    
Hash: 77299979c1a9afc21683
Version: webpack 2.3.3
Time: 2808ms

did it: 1298.503ms                                                    
Hash: fd31795dd4d692c70ccd
Version: webpack 2.3.3
Time: 3032ms
```

## After Fixing Memory Leak

```
did it: 252.222ms                                                     
Hash: fd31795dd4d692c70ccd
Version: webpack 2.3.3
Time: 2157ms

did it: 220.383ms                                                     
Hash: 77299979c1a9afc21683
Version: webpack 2.3.3
Time: 1999ms

did it: 236.641ms                                                     
Hash: fd31795dd4d692c70ccd
Version: webpack 2.3.3
```

## After Fixing Memory Leak + Switching to Set()

```
did it: 3.965ms                                                       
Hash: 77299979c1a9afc21683
Version: webpack 2.3.3
Time: 1949ms

did it: 2.391ms                                                       
Hash: fd31795dd4d692c70ccd
Version: webpack 2.3.3
Time: 1850ms
```

Let me know if you have any questions. Thanks!